Overwriting shell scripts is bad...
During the execution of the init.sh, it would copy the old, unmodified
init.sh from the server folder. This deletes all the custom variables we
have set in the script.

Solve it by using a "start.sh" script that gets generated each time and
doesn't exist in the server directory