External servers part 1
External servers are those separate from MineCloud (i.e., they are not
launched dynamically), but should still be connected to the network. For
example: a persistent Creative server linked to a Minigames server.

These servers should be added to Bungee and counted in the playercount,
but should never be launched or shut down. They also DO NOT run the
MineCloud bukkit plugin.