Change some implementations to fit a minigames server better
NOTE:
This marks the end of me matching style and ideology with the original
MineCloud project. From here on out, the implementation of this system
will be designed with running a Minigames network in mind. Previously
abstracted ideas will be made more concrete and new features will be
added that make running a minigames network easier.

Given that MineCloud is basically dead, I'm not going to try and keep
changes compatible. Their last commit was over 6 months ago... 

Welcome to the brave new world of Cardinal!